# frnd/openapi2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TenderHack Backend",
    "description": "Backend template for a distributed monolith architecture.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth",
          "auth"
        ],
        "summary": "Authenticate user and issue access token",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth",
          "auth"
        ],
        "summary": "Return current authenticated user",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/cte/{cte_id}": {
      "get": {
        "tags": [
          "cte",
          "cte"
        ],
        "summary": "Get CTE by ID",
        "operationId": "get_cte_by_id_api_v1_cte__cte_id__get",
        "parameters": [
          {
            "name": "cte_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cte Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CteItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cte/organizations": {
      "post": {
        "tags": [
          "cte",
          "cte"
        ],
        "summary": "Get organizations by CTE IDs",
        "description": "Return unique organizations (buyers and suppliers) linked to the given cte_ids\nvia contract_items → contracts → organizations.",
        "operationId": "get_organizations_by_cte_ids_api_v1_cte_organizations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CteOrganizationsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CteOrganizationsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/nmck/{cte_id}": {
      "get": {
        "tags": [
          "nmck",
          "nmck"
        ],
        "summary": "Calculate НМЦК via vector search + contracts",
        "operationId": "calculate_nmck_by_search_api_v1_nmck__cte_id__get",
        "parameters": [
          {
            "name": "cte_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cte Id"
            }
          },
          {
            "name": "top_n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of similar CTEs from Qdrant",
              "default": 5,
              "title": "Top N"
            },
            "description": "Number of similar CTEs from Qdrant"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Contract signed_at lower bound",
              "title": "Date From"
            },
            "description": "Contract signed_at lower bound"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Contract signed_at upper bound",
              "title": "Date To"
            },
            "description": "Contract signed_at upper bound"
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Buyer region filter (substring)",
              "title": "Region"
            },
            "description": "Buyer region filter (substring)"
          },
          {
            "name": "inn",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Organization INN filter",
              "title": "Inn"
            },
            "description": "Organization INN filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchNmckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/nmck/manual": {
      "post": {
        "tags": [
          "nmck",
          "nmck"
        ],
        "summary": "Calculate НМЦК from user-provided contracts",
        "description": "User provides their own contracts with prices for a given cte_id.\nНМЦК is computed purely from those prices.",
        "operationId": "calculate_nmck_manual_api_v1_nmck_manual_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualNmckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualNmckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "schemas": {
      "ContractItemOut": {
        "properties": {
          "cte_id": {
            "type": "integer",
            "title": "Cte Id"
          },
          "cte_position_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cte Position Name"
          },
          "unit_price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Price"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "unit_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Name"
          }
        },
        "type": "object",
        "required": [
          "cte_id",
          "cte_position_name",
          "unit_price",
          "quantity",
          "unit_name"
        ],
        "title": "ContractItemOut"
      },
      "ContractOut": {
        "properties": {
          "contract_id": {
            "type": "integer",
            "title": "Contract Id"
          },
          "contract_external_id": {
            "type": "integer",
            "title": "Contract External Id"
          },
          "purchase_name": {
            "type": "string",
            "title": "Purchase Name"
          },
          "signed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed At"
          },
          "buyer_inn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buyer Inn"
          },
          "buyer_region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buyer Region"
          },
          "supplier_inn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Inn"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ContractItemOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "contract_id",
          "contract_external_id",
          "purchase_name",
          "signed_at",
          "buyer_inn",
          "buyer_region",
          "supplier_inn",
          "items"
        ],
        "title": "ContractOut"
      },
      "CteItem": {
        "properties": {
          "cte_id": {
            "type": "string",
            "title": "Cte Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "manufacturer": {
            "type": "string",
            "title": "Manufacturer"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "price_normalized": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Normalized"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "characteristics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Characteristics"
          }
        },
        "type": "object",
        "required": [
          "cte_id",
          "category",
          "manufacturer",
          "description",
          "price",
          "price_normalized",
          "region",
          "characteristics"
        ],
        "title": "CteItem"
      },
...
      "CteOrganizationsRequest": {
        "properties": {
          "cte_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "minItems": 1,
            "title": "Cte Ids",
            "description": "List of CTE IDs to look up"
          }
        },
        "type": "object",
        "required": [
          "cte_ids"
        ],
        "title": "CteOrganizationsRequest"
      },
      "CteOrganizationsResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "organizations": {
            "items": {
              "$ref": "#/components/schemas/OrganizationResponse"
            },
            "type": "array",
            "title": "Organizations"
          }
        },
        "type": "object",
        "required": [
          "total",
          "organizations"
        ],
        "title": "CteOrganizationsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "title": "Username",
            "examples": [
              "demo"
            ]
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password",
            "examples": [
              "demo-password"
            ]
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
      "ManualContractInput": {
        "properties": {
          "unit_price": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Unit Price",
            "description": "Price per unit"
          },
          "signed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed At",
            "description": "Contract date"
          },
          "supplier_inn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Inn",
            "description": "Supplier INN"
          }
        },
        "type": "object",
        "required": [
          "unit_price"
        ],
        "title": "ManualContractInput"
      },
      "ManualNmckRequest": {
        "properties": {
          "cte_id": {
            "type": "integer",
            "title": "Cte Id"
          },
          "contracts": {
            "items": {
              "$ref": "#/components/schemas/ManualContractInput"
            },
            "type": "array",
            "minItems": 1,
            "title": "Contracts",
            "description": "User-provided contracts"
          }
        },
        "type": "object",
        "required": [
          "cte_id",
          "contracts"
        ],
        "title": "ManualNmckRequest"
      },
      "ManualNmckResponse": {
        "properties": {
          "cte_id": {
            "type": "integer",
            "title": "Cte Id"
          },
          "cte_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cte Name"
          },
          "contracts": {
            "items": {
              "$ref": "#/components/schemas/ManualContractInput"
            },
            "type": "array",
            "title": "Contracts"
          },
          "nmck_per_unit": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Nmck Per Unit"
          },
          "price_min": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price Min"
          },
          "price_max": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price Max"
          },
          "mean_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Mean Price"
          },
          "median_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Median Price"
          },
          "std_dev": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Std Dev"
          },
          "coefficient_of_variation": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Coefficient Of Variation"
          },
          "is_homogeneous": {
            "type": "boolean",
            "title": "Is Homogeneous"
          },
          "confidence_interval_low": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confidence Interval Low"
          },
          "confidence_interval_high": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confidence Interval High"
          },
          "sample_size_total": {
            "type": "integer",
            "title": "Sample Size Total"
          },
          "sample_size_filtered": {
            "type": "integer",
            "title": "Sample Size Filtered"
          },
          "outliers_removed": {
            "type": "integer",
            "title": "Outliers Removed"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "method": {
            "type": "string",
            "title": "Method",
            "default": "comparable_market_prices"
          }
        },
        "type": "object",
        "required": [
          "cte_id",
          "cte_name",
          "contracts",
          "nmck_per_unit",
          "price_min",
          "price_max",
          "mean_price",
          "median_price",
          "std_dev",
          "coefficient_of_variation",
          "is_homogeneous",
          "confidence_interval_low",
          "confidence_interval_high",
          "sample_size_total",
          "sample_size_filtered",
          "outliers_removed",
          "warnings"
        ],
        "title": "ManualNmckResponse"
      },
      "OrganizationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "inn": {
            "type": "string",
            "title": "Inn"
          },
          "region_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Name"
          },
          "org_type": {
            "type": "string",
            "title": "Org Type"
          }
        },
        "type": "object",
        "required": [
          "id",
          "inn",
          "region_name",
          "org_type"
        ],
        "title": "OrganizationResponse"
      },
...
      "SearchNmckResponse": {
        "properties": {
          "cte_id": {
            "type": "integer",
            "title": "Cte Id"
          },
          "cte_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cte Name"
          },
          "search_query": {
            "type": "string",
            "title": "Search Query"
          },
          "matched_cte_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Matched Cte Ids"
          },
          "matched_cte_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Cte Names"
          },
          "prices_by_cte": {
            "additionalProperties": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Prices By Cte"
          },
          "contracts": {
            "items": {
              "$ref": "#/components/schemas/ContractOut"
            },
            "type": "array",
            "title": "Contracts"
          },
          "total_contracts": {
            "type": "integer",
            "title": "Total Contracts"
          },
          "nmck_per_unit": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Nmck Per Unit"
          },
          "price_min": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price Min"
          },
          "price_max": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price Max"
          },
          "mean_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Mean Price"
          },
          "median_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Median Price"
          },
          "std_dev": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Std Dev"
          },
          "coefficient_of_variation": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Coefficient Of Variation"
          },
          "is_homogeneous": {
            "type": "boolean",
            "title": "Is Homogeneous"
          },
          "confidence_interval_low": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confidence Interval Low"
          },
          "confidence_interval_high": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confidence Interval High"
          },
          "sample_size_total": {
            "type": "integer",
            "title": "Sample Size Total"
          },
          "sample_size_filtered": {
            "type": "integer",
            "title": "Sample Size Filtered"
          },
          "outliers_removed": {
            "type": "integer",
            "title": "Outliers Removed"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "method": {
            "type": "string",
            "title": "Method",
            "default": "comparable_market_prices"
          }
        },
        "type": "object",
        "required": [
          "cte_id",
          "cte_name",
          "search_query",
          "matched_cte_ids",
          "matched_cte_names",
          "prices_by_cte",
          "contracts",
          "total_contracts",
          "nmck_per_unit",
          "price_min",
          "price_max",
          "mean_price",
          "median_price",
          "std_dev",
          "coefficient_of_variation",
          "is_homogeneous",
          "confidence_interval_low",
          "confidence_interval_high",
          "sample_size_total",
          "sample_size_filtered",
          "outliers_removed",
          "warnings"
        ],
        "title": "SearchNmckResponse"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenResponse"
      },
      "UserResponse": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "username"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}